+++ test/unit/index.ts
@@ -6,26 +6,24 @@
  * Copyright (c) 2014 Charlike Mike Reagent, Daryl Lau, contributors.
  * Released under the MIT license.
  */
+import assert from 'assert';
+import fs from 'fs';
+import http from 'http';
+import koaBody from '../../src/index';
+import path from 'path';
+import request, { Response } from 'supertest';
+import should from 'should';
+import Koa from 'koa';
+import Router from 'koa-router';
+import sinon, { SinonSpy } from 'sinon';
 
-'use strict';
-
-const assert = require('assert');
-const fs = require('fs');
-const http = require('http');
-const koaBody = require('../../index.js');
-const path = require('path');
-const request = require('supertest');
-const should = require('should');
-const Koa = require('koa');
-const Router = require('koa-router');
-const sinon = require('sinon');
-
-const unparsed = require('../../unparsed.js');
+import unparsed from '../../src/unparsed';
+import { HttpMethodEnum } from '../../src/types';
 
 describe('koa-body', () => {
-  let database;
-  let router;
-  let app;
+  let database: { users: Array<{ name: string; followers?: number }> };
+  let router: Router;
+  let app: Koa;
 
   beforeEach((done) => {
     app = new Koa();
@@ -33,33 +31,33 @@ describe('koa-body', () => {
       users: [
         {
           name: 'charlike',
-          followers: 10
+          followers: 10,
         },
         {
           name: 'tunnckocore',
-          followers: 20
-        }
-      ]
+          followers: 20,
+        },
+      ],
     };
-    router = Router()
+    router = new Router()
       .get('/users', (ctx, next) => {
-        if(ctx.request.body && ctx.request.body.name) {
-          ctx.body = database.users.find(element => element.name === ctx.request.body.name);
+        if (ctx.request.body && ctx.request.body.name) {
+          ctx.body = database.users.find((element) => element.name === ctx.request.body.name);
           ctx.status = 200;
           return next();
         }
         ctx.status = 200;
         ctx.body = database;
       })
-      .get('/users/:user', (ctx, next) => {
-        user = database.users.find(element => element.name === ctx.request.body.name);
+      .get('/users/:user', (ctx) => {
+        const user = database.users.find((element) => element.name === ctx.request.body.name);
         ctx.status = 200;
         ctx.body = user;
       })
       .post('/users', (ctx, next) => {
         const user = ctx.request.body;
 
-        if(!user) {
+        if (!user) {
           ctx.status = 400;
           return next();
         }
@@ -69,21 +67,21 @@ describe('koa-body', () => {
         // return request contents to validate
         ctx.body = {
           _files: ctx.request.files, // files we populate
-          user: user // original request data
+          user: user, // original request data
         };
       })
-      .post('/echo_body', (ctx, next) => {
+      .post('/echo_body', (ctx) => {
         ctx.status = 200;
         ctx.body = ctx.request.body;
       })
-      .delete('/users/:user', (ctx, next) => {
+      .delete('/users/:user', (ctx) => {
         const user = ctx.params.user;
         const multi = !!ctx.request.body.multi;
         if (multi) {
-          database.users = database.users.filter(element => element.name !== user);
-        }
-        else {
-          const index = database.users.findIndex(element => element === user);
+          database.users = database.users.filter((element) => element.name !== user);
+        } else {
+          // @ts-ignore
+          const index = database.users.findIndex((element) => element === user);
           database.users.splice(index, 1);
         }
         ctx.status = 204;
@@ -92,28 +90,26 @@ describe('koa-body', () => {
     done();
   });
 
-
   /**
    * DEFAULTS - multipart: false
    */
-  it('should work with defaults - multipart: false, only `urlencoded` and `json` bodies',  (done) => {
+  it('should work with defaults - multipart: false, only `urlencoded` and `json` bodies', (done) => {
     app.use(koaBody());
     app.use(router.routes());
 
     request(http.createServer(app.callback()))
       .get('/users')
       .expect(200, database)
-      .end( (err, res) => {
+      .end((err) => {
         if (err) return done(err);
         done();
       });
   });
 
-
   /**
    * MULTIPART - FIELDS
    */
-  it('should receive `multipart` requests - fields on .body object',  (done) => {
+  it('should receive `multipart` requests - fields on .body object', (done) => {
     app.use(koaBody({ multipart: true }));
     app.use(router.routes());
 
@@ -122,7 +118,7 @@ describe('koa-body', () => {
       .field('name', 'daryl')
       .field('followers', 30)
       .expect(201)
-      .end( (err, res) => {
+      .end((err, res) => {
         if (err) return done(err);
 
         const mostRecentUser = database.users[database.users.length - 1];
@@ -137,17 +133,18 @@ describe('koa-body', () => {
       });
   });
 
-
   /**
    * MULTIPART - FILES
    */
-  it('should receive multiple fields and files via `multipart` on .body.files object',  (done) => {
-    app.use(koaBody({
-      multipart: true,
-      formidable: {
-        uploadDir: '/tmp'
-      }
-    }));
+  it('should receive multiple fields and files via `multipart` on .body.files object', (done) => {
+    app.use(
+      koaBody({
+        multipart: true,
+        formidable: {
+          uploadDir: '/tmp',
+        },
+      }),
+    );
     app.use(router.routes());
 
     request(http.createServer(app.callback()))
@@ -156,13 +153,13 @@ describe('koa-body', () => {
       .field('names', 'John')
       .field('names', 'Paul')
       .attach('firstField', 'package.json')
-      .attach('secondField', 'index.js')
+      .attach('secondField', 'src/index.ts')
       .attach('secondField', 'package.json')
       .attach('thirdField', 'LICENSE')
       .attach('thirdField', 'README.md')
       .attach('thirdField', 'package.json')
       .expect(201)
-      .end( (err, res) => {
+      .end((err, res) => {
         if (err) return done(err);
         res.body.user.names.should.be.an.Array().and.have.lengthOf(2);
         res.body.user.names[0].should.equal('John');
@@ -173,12 +170,16 @@ describe('koa-body', () => {
         fs.unlinkSync(res.body._files.firstField.filepath);
 
         res.body._files.secondField.should.be.an.Array().and.have.lengthOf(2);
-        res.body._files.secondField.should.containDeep([{
-          originalFilename: 'index.js'
-        }]);
-        res.body._files.secondField.should.containDeep([{
-          originalFilename: 'package.json'
-        }]);
+        res.body._files.secondField.should.containDeep([
+          {
+            originalFilename: 'index.ts',
+          },
+        ]);
+        res.body._files.secondField.should.containDeep([
+          {
+            originalFilename: 'package.json',
+          },
+        ]);
         should(fs.statSync(res.body._files.secondField[0].filepath)).be.ok();
         should(fs.statSync(res.body._files.secondField[1].filepath)).be.ok();
         fs.unlinkSync(res.body._files.secondField[0].filepath);
@@ -186,15 +187,21 @@ describe('koa-body', () => {
 
         res.body._files.thirdField.should.be.an.Array().and.have.lengthOf(3);
 
-        res.body._files.thirdField.should.containDeep([{
-          originalFilename: 'LICENSE'
-        }]);
-        res.body._files.thirdField.should.containDeep([{
-          originalFilename: 'README.md'
-        }]);
-        res.body._files.thirdField.should.containDeep([{
-          originalFilename: 'package.json'
-        }]);
+        res.body._files.thirdField.should.containDeep([
+          {
+            originalFilename: 'LICENSE',
+          },
+        ]);
+        res.body._files.thirdField.should.containDeep([
+          {
+            originalFilename: 'README.md',
+          },
+        ]);
+        res.body._files.thirdField.should.containDeep([
+          {
+            originalFilename: 'package.json',
+          },
+        ]);
         should(fs.statSync(res.body._files.thirdField[0].filepath)).be.ok();
         fs.unlinkSync(res.body._files.thirdField[0].filepath);
         should(fs.statSync(res.body._files.thirdField[1].filepath)).be.ok();
@@ -206,18 +213,20 @@ describe('koa-body', () => {
       });
   });
 
-  it('can transform file names in multipart requests',  (done) => {
-    app.use(koaBody({
-      multipart: true,
-      formidable: {
-        uploadDir: '/tmp',
-        onFileBegin:  (name, file) => {
-          file.newFilename = 'backage.json'
-          const folder = path.dirname(file.filepath);
-          file.filepath = path.join(folder, file.newFilename);
-        }
-      }
-    }));
+  it('can transform file names in multipart requests', (done) => {
+    app.use(
+      koaBody({
+        multipart: true,
+        formidable: {
+          uploadDir: '/tmp',
+          onFileBegin: (_name, file) => {
+            file.newFilename = 'backage.json';
+            const folder = path.dirname(file.filepath);
+            file.filepath = path.join(folder, file.newFilename);
+          },
+        },
+      }),
+    );
     app.use(router.routes());
 
     request(http.createServer(app.callback()))
@@ -225,7 +234,7 @@ describe('koa-body', () => {
       .type('multipart/form-data')
       .attach('firstField', 'package.json')
       .expect(201)
-      .end( (err, res) => {
+      .end((err, res) => {
         if (err) return done(err);
 
         res.body._files.firstField.should.be.an.Object();
@@ -237,11 +246,10 @@ describe('koa-body', () => {
       });
   });
 
-
   /**
    * URLENCODED request body
    */
-  it('should recieve `urlencoded` request bodies',  (done) => {
+  it('should recieve `urlencoded` request bodies', (done) => {
     app.use(koaBody({ multipart: true }));
     app.use(router.routes());
 
@@ -250,10 +258,10 @@ describe('koa-body', () => {
       .type('application/x-www-form-urlencoded')
       .send({
         name: 'example',
-        followers: '41'
+        followers: '41',
       })
       .expect(201)
-      .end( (err, res) => {
+      .end((err, res) => {
         if (err) return done(err);
 
         const mostRecentUser = database.users[database.users.length - 1];
@@ -269,8 +277,7 @@ describe('koa-body', () => {
    * Inclusion of unparsed body when opts.includeUnparsed is true
    */
   describe('includeUnparsed tests', () => {
-
-    let requestSpy;
+    let requestSpy: SinonSpy | undefined;
 
     beforeEach(() => {
       app.use(koaBody({ includeUnparsed: true }));
@@ -278,12 +285,13 @@ describe('koa-body', () => {
     });
 
     afterEach(() => {
-      requestSpy.restore();
+      requestSpy?.restore();
       requestSpy = undefined;
     });
 
     it('should not fail when no request body is provided', (done) => {
-      const echoRouterLayer = router.stack.filter(layer => layer.path === "/echo_body");
+      const echoRouterLayer = router.stack.filter((layer) => layer.path === '/echo_body');
+      // @ts-ignore
       requestSpy = sinon.spy(echoRouterLayer[0].stack, '0');
 
       request(http.createServer(app.callback()))
@@ -294,8 +302,8 @@ describe('koa-body', () => {
         .end((err, res) => {
           if (err) return done(err);
 
-          assert(requestSpy.calledOnce, 'Spy for /echo_body not called');
-          const req = requestSpy.firstCall.args[0].request;
+          assert(requestSpy?.calledOnce, 'Spy for /echo_body not called');
+          const req = requestSpy?.firstCall.args[0].request;
 
           req.body.should.not.be.Undefined();
           res.body.should.have.properties({});
@@ -308,9 +316,11 @@ describe('koa-body', () => {
         });
     });
 
-    it('should recieve `urlencoded` request bodies with the `includeUnparsed` option',  (done) => {
-
-      const userRouterLayer = router.stack.filter(layer => layer.path === "/users" && layer.methods.includes('POST'));
+    it('should recieve `urlencoded` request bodies with the `includeUnparsed` option', (done) => {
+      const userRouterLayer = router.stack.filter(
+        (layer) => layer.path === '/users' && layer.methods.includes('POST'),
+      );
+      // @ts-ignore
       requestSpy = sinon.spy(userRouterLayer[0].stack, '0');
 
       request(http.createServer(app.callback()))
@@ -318,16 +328,16 @@ describe('koa-body', () => {
         .type('application/x-www-form-urlencoded')
         .send({
           name: 'Test',
-          followers: '97'
+          followers: '97',
         })
         .expect(201)
-        .end( (err, res) => {
+        .end((err, res) => {
           if (err) return done(err);
 
           const mostRecentUser = database.users[database.users.length - 1];
 
-          assert(requestSpy.calledOnce, 'Spy for /users not called');
-          const req = requestSpy.firstCall.args[0].request;
+          assert(requestSpy?.calledOnce, 'Spy for /users not called');
+          const req = requestSpy?.firstCall.args[0].request;
           req.body[unparsed].should.not.be.Undefined();
           req.body[unparsed].should.be.a.String();
           req.body[unparsed].should.equal('name=Test&followers=97');
@@ -339,8 +349,8 @@ describe('koa-body', () => {
     });
 
     it('should receive JSON request bodies as strings with the `includeUnparsed` option', (done) => {
-
-      const echoRouterLayer = router.stack.filter(layer => layer.path === "/echo_body");
+      const echoRouterLayer = router.stack.filter((layer) => layer.path === '/echo_body');
+      // @ts-ignore
       requestSpy = sinon.spy(echoRouterLayer[0].stack, '0');
 
       request(http.createServer(app.callback()))
@@ -348,20 +358,22 @@ describe('koa-body', () => {
         .type('application/json')
         .send({
           hello: 'world',
-          number: 42
+          number: 42,
         })
         .expect(200)
         .end((err, res) => {
           if (err) return done(err);
 
-          assert(requestSpy.calledOnce, 'Spy for /echo_body not called');
-          const req = requestSpy.firstCall.args[0].request;
+          assert(requestSpy?.calledOnce, 'Spy for /echo_body not called');
+          const req = requestSpy?.firstCall.args[0].request;
           req.body[unparsed].should.not.be.Undefined();
           req.body[unparsed].should.be.a.String();
-          req.body[unparsed].should.equal(JSON.stringify({
-            hello: 'world',
-            number: 42
-          }));
+          req.body[unparsed].should.equal(
+            JSON.stringify({
+              hello: 'world',
+              number: 42,
+            }),
+          );
 
           res.body.should.have.properties({ hello: 'world', number: 42 });
 
@@ -370,8 +382,8 @@ describe('koa-body', () => {
     });
 
     it('should receive text as strings with `includeUnparsed` option', (done) => {
-
-      const echoRouterLayer = router.stack.filter(layer => layer.path === "/echo_body");
+      const echoRouterLayer = router.stack.filter((layer) => layer.path === '/echo_body');
+      // @ts-ignore
       requestSpy = sinon.spy(echoRouterLayer[0].stack, '0');
 
       request(http.createServer(app.callback()))
@@ -382,8 +394,8 @@ describe('koa-body', () => {
         .end((err, res) => {
           if (err) return done(err);
 
-          assert(requestSpy.calledOnce, 'Spy for /echo_body not called');
-          const req = requestSpy.firstCall.args[0].request;
+          assert(requestSpy?.calledOnce, 'Spy for /echo_body not called');
+          const req = requestSpy?.firstCall.args[0].request;
           should(req.body).equal('plain text content');
 
           // Raw text requests are still just text
@@ -396,7 +408,6 @@ describe('koa-body', () => {
           done();
         });
     });
-
   });
 
   /**
@@ -411,7 +422,7 @@ describe('koa-body', () => {
       .type('text')
       .send('plain text')
       .expect(200)
-      .end( (err, res) => {
+      .end((err, res) => {
         if (err) return done(err);
 
         res.type.should.equal('text/plain');
@@ -426,7 +437,8 @@ describe('koa-body', () => {
     app.use(router.routes());
     const body = '<?xml version="1.0"?><catalog></catalog>';
 
-    const echoRouterLayer = router.stack.filter(layer => layer.path === "/echo_body");
+    const echoRouterLayer = router.stack.filter((layer) => layer.path === '/echo_body');
+    // @ts-ignore
     const requestSpy = sinon.spy(echoRouterLayer[0].stack, '0');
 
     request(http.createServer(app.callback()))
@@ -434,11 +446,11 @@ describe('koa-body', () => {
       .type('text/xml')
       .send(body)
       .expect(200)
-      .end( (err, res) => {
+      .end((err, res) => {
         if (err) return done(err);
 
-        assert(requestSpy.calledOnce, 'Spy for /echo_body not called');
-        const req = requestSpy.firstCall.args[0].request;
+        assert(requestSpy?.calledOnce, 'Spy for /echo_body not called');
+        const req = requestSpy?.firstCall.args[0].request;
         assert.equal(req.body[unparsed], undefined);
         req.body.should.equal(body);
         res.text.should.equal(body);
@@ -447,14 +459,23 @@ describe('koa-body', () => {
       });
   });
 
-  describe('strict mode',  () => {
-    beforeEach( () => {
+  describe('parsedMethods options', () => {
+    beforeEach(() => {
       //push an additional, to test the multi query
       database.users.push({ name: 'charlike' });
     });
 
-    it('can enable strict mode',  (done) => {
-      app.use(koaBody({ strict: true }));
+    it('methods declared are parsed', (done) => {
+      app.use(
+        koaBody({
+          parsedMethods: [
+            HttpMethodEnum.PATCH,
+            HttpMethodEnum.POST,
+            HttpMethodEnum.PUT,
+            HttpMethodEnum.DELETE,
+          ],
+        }),
+      );
       app.use(router.routes());
 
       request(http.createServer(app.callback()))
@@ -462,15 +483,17 @@ describe('koa-body', () => {
         .type('application/x-www-form-urlencoded')
         .send({ multi: true })
         .expect(204)
-        .end( (err, res) => {
+        .end((err) => {
           if (err) return done(err);
-          assert(database.users.find(element => element.name === 'charlike') !== undefined);
+          assert(database.users.find((element) => element.name === 'charlike') === undefined);
           done();
         });
     });
 
-    it('can disable strict mode',  (done) => {
-      app.use(koaBody({ strict: false }));
+    it('methods do not get parsed if not declared', (done) => {
+      app.use(
+        koaBody({ parsedMethods: [HttpMethodEnum.POST, HttpMethodEnum.PUT, HttpMethodEnum.PATCH] }),
+      );
       app.use(router.routes());
 
       request(http.createServer(app.callback()))
@@ -478,76 +501,20 @@ describe('koa-body', () => {
         .type('application/x-www-form-urlencoded')
         .send({ multi: true })
         .expect(204)
-        .end( (err, res) => {
+        .end((err) => {
           if (err) return done(err);
-          assert(database.users.find(element => element.name === 'charlike') === undefined);
+          assert(database.users.find((element) => element.name === 'charlike') !== undefined);
           done();
         });
-    });
-  });
-
-  describe('parsedMethods options',  () => {
-    beforeEach( () => {
-      //push an additional, to test the multi query
-      database.users.push({ name: 'charlike' });
-    });
-
-    it('methods declared are parsed',  (done) => {
-      app.use(koaBody({ parsedMethods: ['POST', 'PUT', 'PATCH', 'DELETE'] }));
-      app.use(router.routes());
-
-      request(http.createServer(app.callback()))
-        .delete('/users/charlike')
-        .type('application/x-www-form-urlencoded')
-        .send({ multi: true })
-        .expect(204)
-        .end( (err, res) => {
-          if (err) return done(err);
-          assert(database.users.find(element => element.name === 'charlike') === undefined);
-          done();
-        });
-    });
-
-    it('methods do not get parsed if not declared',  (done) => {
-      app.use(koaBody({ parsedMethods: ['POST', 'PUT', 'PATCH'] }));
-      app.use(router.routes());
-
-      request(http.createServer(app.callback()))
-        .delete('/users/charlike')
-        .type('application/x-www-form-urlencoded')
-        .send({ multi: true })
-        .expect(204)
-        .end( (err, res) => {
-          if (err) return done(err);
-          assert(database.users.find(element => element.name === 'charlike') !== undefined);
-          done();
-        });
-    });
-
-    it('cannot use strict mode and parsedMethods options at the same time', (done) => {
-      let err;
-      try {
-        app.use(koaBody({
-          parsedMethods: ['POST', 'PUT', 'PATCH'],
-          strict: true
-        }));
-      } catch (_err) {
-        err = _err;
-      }
-
-      assert(err && err.message === 'Cannot use strict and parsedMethods options at the same time.');
-
-      done();
     });
   });
 
   /**
    * JSON request body
    */
-  describe('POST json request body',  () => {
-
-    it('should set the follower count',  (done) => {
-      app.use(koaBody({ strict: false }));
+  describe('POST json request body', () => {
+    it('should set the follower count', (done) => {
+      app.use(koaBody());
       app.use(router.routes());
 
       request(http.createServer(app.callback()))
@@ -555,42 +522,46 @@ describe('koa-body', () => {
         .type('application/json')
         .send({
           name: 'json',
-          followers: '313'
+          followers: '313',
         })
         .expect(201)
         .end((err, res) => {
-          res.body.user.should.have.properties({ followers: "313", name: "json" });
+          res.body.user.should.have.properties({ followers: '313', name: 'json' });
           done(err);
         });
     });
   });
 
   describe('GET json request body', () => {
-    let response;
+    let response: Response;
 
     beforeEach((done) => {
-      app.use(koaBody({ strict: false }));
+      app.use(
+        koaBody({
+          parsedMethods: [HttpMethodEnum.GET],
+        }),
+      );
       app.use(router.routes());
       database.users.push({
         name: 'foo',
-        followers: 111
+        followers: 111,
       });
       request(http.createServer(app.callback()))
         .get('/users')
         .type('application/json')
         .send({ name: 'foo' })
         .expect(200)
-        .end( (err, res) => {
+        .end((err, res) => {
           response = res;
           done(err);
-        })
+        });
     });
 
-    it('should parse the response body',  () => {
+    it('should parse the response body', () => {
       response.body.should.not.equal(null);
     });
 
-    it('should return the user details',  () => {
+    it('should return the user details', () => {
       response.body.name.should.equal('foo');
       response.body.followers.should.equal(111);
     });
@@ -617,12 +588,16 @@ describe('koa-body', () => {
           .type(type)
           .send({
             a: 'foo',
-            b: [42]
+            b: [42],
           })
-          .expect(200, {
-            a: 'foo',
-            b: [42]
-          }, done);
+          .expect(
+            200,
+            {
+              a: 'foo',
+              b: [42],
+            },
+            done,
+          );
       });
     }
 
@@ -634,14 +609,22 @@ describe('koa-body', () => {
         request(http.createServer(app.callback()))
           .post('/echo_body')
           .type(type)
-          .send([{
-            a: 'foo',
-            b: [42]
-          }])
-          .expect(200, [{
-            a: 'foo',
-            b: [42]
-          }], done);
+          .send([
+            {
+              a: 'foo',
+              b: [42],
+            },
+          ])
+          .expect(
+            200,
+            [
+              {
+                a: 'foo',
+                b: [42],
+              },
+            ],
+            done,
+          );
       });
     }
   });
@@ -651,7 +634,7 @@ describe('koa-body', () => {
   /**
    * FORM (urlencoded) LIMIT
    */
-  it('should request 413 '+ERR_413_STATUSTEXT+', because of `formLimit`',  (done) => {
+  it('should request 413 ' + ERR_413_STATUSTEXT + ', because of `formLimit`', (done) => {
     app.use(koaBody({ formLimit: 10 /*bytes*/ }));
     app.use(router.routes());
 
@@ -663,11 +646,10 @@ describe('koa-body', () => {
       .end(done);
   });
 
-
   /**
    * JSON LIMIT
    */
-  it('should request 413 '+ERR_413_STATUSTEXT+', because of `jsonLimit`',  (done) => {
+  it('should request 413 ' + ERR_413_STATUSTEXT + ', because of `jsonLimit`', (done) => {
     app.use(koaBody({ jsonLimit: 10 /*bytes*/ }));
     app.use(router.routes());
 
@@ -679,8 +661,7 @@ describe('koa-body', () => {
       .end(done);
   });
 
-
-  it('should tolerate no content type',  (done) => {
+  it('should tolerate no content type', (done) => {
     app.use(koaBody());
     app.use(router.routes());
 
@@ -691,11 +672,10 @@ describe('koa-body', () => {
       .end(done);
   });
 
-
   /**
    * TEXT LIMIT
    */
-  it('should request 413 '+ERR_413_STATUSTEXT+', because of `textLimit`',  (done) =>  {
+  it('should request 413 ' + ERR_413_STATUSTEXT + ', because of `textLimit`', (done) => {
     app.use(koaBody({ textLimit: 10 /*bytes*/ }));
     app.use(router.routes());